introducing p2p pkg
I think it's time to move a lot of the peer-to-peer networking
but-not-ipfs-specific things into its own package: p2p.
This could in the future be split off into its own library.
The first thing to go is the peer.


This commit was moved from ipfs/go-ipfs-routing@888ed12fbc61fcc99e49620d6074f38b99b3ed7e