merkledag: LONG timeout on Get
we shouldn't use an arbitrary timeout here. since Get
doesnt take in a context yet, we give a large upper bound.
think of an http request. we want it to go on as long as
the client requests it.


This commit was moved from ipfs/go-merkledag@339d2a983d4e0fa4b39a308373afebf11ccbd64d