p2p/net: notify on listens
Network now signals when it successfully listens on some address
or when an address shuts down. This will be used to establish and
close nat port mappings. It could also be used to notify peers
of address changes.


This commit was moved from ipfs/go-bitswap@2027fbe0138ea24d2bdbd4738bca64adcca6fce4