corehttp: ensure node closing/teardown waits for server termination
When closing a node, the node itself only takes care of tearing down
its own children. As corehttp sets up a server based on a node, it
needs to also ensure that the server is accounted for when determining
if the node has been fully closed.


This commit was moved from ipfs/kubo@c9d308491004ae42dc7757c6f5176846410876de