Localize the scope of context.WithCancel for every DAG.Get
Instead put it inside of DAG.Get.
The fix is applied only in the case when the context.WithCancel
before a DAG.Get is also used later on in the scope.

License: MIT
Signed-off-by: rht <rhtbot@gmail.com>


This commit was moved from ipfs/kubo@de5c0ceff0857ee3ee4be85785b3253c0ef98c28