Add Suborigin header to gateway responses (#3209)
This existed before but was disabled in 630596a because the Suborigin spec
changed and it became incompatible.

This commit updates the generated Suborigin header to be conformant with the
latest spec.

License: MIT
Signed-off-by: James Stanley <james@incoherency.co.uk>


This commit was moved from ipfs/kubo@cb2a38d89b3c02b509f1d974b60f114cf7cf44ac