preallocate Batch's blocks buffer on commit.
It's probably safe to assume that this buffer will be about the same time each
flush.

This could cause 1 extra allocation (if this is the last commit) but that's
unlikely to be an issue.

License: MIT
Signed-off-by: Steven Allen <steven@stebalien.com>


This commit was moved from ipfs/go-merkledag@0a5f16ec54dcad7cf4d51c7807928794cade0499