show hash if not in original url
Change the template handler to pass the hash to the directory listing template if the original url doesn't contain the hash. The directory listing template will display the hash in grey under the original url if the hash is passed to it.

License: MIT
Signed-off-by: Jack Loughran <j@ckloughran.com>


This commit was moved from ipfs/kubo@9ef14863075ff30a75d35dea5d1ca7ef2a94e800