session: buffer some request channels
We're not using these synchronously so we can buffer them a bit to avoid
blocking quite as much.

This also combines all incoming channels into a single one to ensure all
operations are processed in-order. This might be overkill bit it makes reasoning
about this a bit simpler.


This commit was moved from ipfs/go-bitswap@8454ba009515209fc7cc74e320a8a03ee993def4