feat: wildcard support for public gateways
Add support for one or more wildcards in the hostname definition
of a public gateway. This is useful for example to support easily
multiples environment.

Wildcarded hostname are set in the config as for example "*.domain.tld".


This commit was moved from ipfs/kubo@13e6bcfb4f7d2db7d80af77f1b32b2394abca9da