fix(gateway): ensure directory listings have Content-Type text/html
Files already have an explicit Content-Type set. Be sure to do this for
directory listings as well to avoid a fallback to autodetection in
net/http. That fallback fails when a ResponseWriter is installed that
performs compression.


This commit was moved from ipfs/kubo@d4952f2a73b2dd33d8c35ab1bae3bc381fdb5088