fix: don't say we're sending a full wantlist unless we are (#429)
I'm not sure why we set "full" to true here, but this could be the source of a
whole bunch of bidirectional sync issues. That is, if two peers are syncing off
each other, they could repeatedly "reset" each other's wantlist to "empty".

This commit was moved from ipfs/go-bitswap@72d351cb3915079401fc3594baab3be50d736650