introduce a ledger that stores which peers are waiting for a Cid
When receiving a new block (Engine.ReceiveFrom), we shouldn't have to loop over
all peers in order to determine if they need this block. Instead, use a map to
save which peers are waiting for a give Cid.


This commit was moved from ipfs/go-bitswap@2bfc771f7941679b9e243477debb68c453e2683e