fix: Nil dereference while using SetSendDontHaves
This option is used by the benchmark to simulate the old bitswap
comportement.

This follows the same refactoring idea as done in 22e70990a3ed.

It was crashing since it was trying to access the `sendDontHaves`
property of `bs.engine` but `bs.engine` is initialized right after
the options are applied, not before.


This commit was moved from ipfs/go-bitswap@f2d9b5a50aee63b0897de8aa8d43052663c0a316