Implement CAR v2 writer that uses `format.NodeGetter`
Implement a CAR v2 writer, that produces binary structure corresponding
to the specification of car V2, consisting of:
 1. Version prefix
 2. CAR v2 header
 3. Dump of Car v1
 4. Carbs index

The implementation also facilities optional padding before dump of CAR
v1 and Carbs index to provide scope for future optimisations. The
padding is defined as a dedicated type that writes zero-valued bytes for
a given padding size. The CAR v2 header, then captures the offsets from
the beginning of the file for both the CAR v1 dump and Carbs index. This
allows the user to quickly skip to the part of the CAR v2 they need as
well as offset alignment, if necessary, by altering the value of
padding.

Note, this is an intermediate implementation, and does not correctly
count the number of bytes written by the writer, pending the
transfer/refactor of Carbs implementation in this repo. In the meantime,
This implementation depends on Carbs as a go module. The implementation
of extensive writer tests is postponed until Carbs is transferred and
the written byte-count can be returned from the index marshaller.

Address review comments

- Avoid representing `Characteristics` as a pointer for easier casting
of memory regions.

- Remove anonymous fields in structs as a human readable way to document
what interfaces a struct implements.

- Simplify Header writing and written byte count calculation. The change
reduces the number of lines by assuming that `writeUint64To` will always
write 8 bytes.

- Avoid `_` in package names. Use `carv1` and `carv2` to name
corresponding packages.

- Rename `Marshal` in tests to reflect `WriteTo` related tests. The
method was renamed in the implementation but tests were not renamed.

- Write padding in bulk to reduce redundant memory allocation. Write
padding in bulks of `1024` bytes to reduce large memory allocation when
padding itself is large.

- Avoid using # in docs. Use go syntax instead, i.e. `.`.

- Remove redundant type in constants, since it is implicitly converted.

- Simplify over-refactored prefix write Since it is called only once.

- Avoid `Header` pointer, since its size is small and this would reduce
unnecessary allocations.

- Use buffer in writer to store encoded Car V1. This is to avoid
reallocation of bytes buffer.

- Add TODO re optimisation of index generation. The current
implementation reads the entire CAR v1 into memory in order to index it,
because `carbs` API requires `io.Reader`. Once `a`carbs` is incorporated
into this repo consider refactoring the API to make this a streaming
operation that avoids copying all the bytes since CAR v1 can be large.

- Remove a dedicated var for empty characteristics, since, we assume the
default to be all zero, and it is easy enough to construct a zero valued
Characteristics.

- Unexport PrefixBytesSize, because we can, and if it is public it might
have to stay public forever. So, unexport until we know we need it
exported.

- Remove `.Size()` on CAR v2 structs. Because we have the constant for
them and don't want to expose them twice.

- Use CamelCase for sub-test naming. This is to establish CamelCase for
sub-test naming as the convention for this repo since it keeps the test
names consistent in code and in the output ot `go test`.

- Unexport padding as a type since it is only used internally


This commit was moved from ipld/go-car@875a6548b3047e46f493f6edd29b2f37d4fb5b4e