Implement `OffsetReader`
Because `SectionReader` requires the max number of bytes to read, since
it implements `Seek`. We need something like the `SectionReader` to read
the index at the end of a CAR v2 that does not require the user to know
the number of readable bytes. This is because, we do not store the size
of index in CAR v2 header, since it is always added as the last section.

The `OffsetReader` works just like `SectionReader`, except if `n`, the
number of bytes to read, is set to zero it simply carries on reading
until the underlying `io.ReaderAt` returns EOF. Consequently,
`OffsetReader` does not implement `Seek`.


This commit was moved from ipld/go-car@7cbf448577b17fa4a1712c5d8f33c99a7cfefea2