Remove dependency to CAR v1 go module
Remove dependency to CAR v1 go module, because that module depends on
later version of ipld-prime which have not been reflected in filecoin
project.

In the interest of time, we fork the code that v2 needs in v1. This
allows a more efficient implementation of some APIs that accep types
like `bufio.Reader` while helping the filecoin team push ahead with
deliverables without undergoing a big upgrade process.

On the CAR v2 side, this however means that we postpone the
implementation of Selective Car for v2 until filecoin is upgraded its
ipld-prime dependency. This is to avoid implementing an obsolete
selective car API that uses the old ipld-prime API which force user yet
another upgrade.


This commit was moved from ipld/go-car@ff99aafeec59dc062b7cb8af6619ae36df1af60e