blockstore: make unsupported methods panic
That is, calling write methods on a read-only blockstore, or any methods
on a finalized blockstore.

The user's code should never be calling these, so there's no reason to
avoid panics. There is one good reason to use panics, though: if
anywhere in a program those rules aren't followed, a panic is much more
obvious and quick to spot than an error which could just get logged or
ignored.

If a user wants a read-only blockstore that returns errors on write
funcs, they could always wrap our read-only blockstore to do that.
Arguably that's their best option anyway, so they are in control of the
exact behavior they need.

Also fix a typo in carV1Wrtier.


This commit was moved from ipld/go-car@c25c2351cfe777e24f8dbbf3432bdbbfcacf83e1